SourceDocument refactoring
