+++ src/main/resources/regression/core17.yaml
@@ -6,7 +6,7 @@ search_command: target/appassembler/bin/SearchCollection
 topic_root: src/main/resources/topics-and-qrels/
 qrels_root: src/main/resources/topics-and-qrels/
 ranking_root:
-generator: JsoupGenerator
+generator: DefaultLuceneDocumentGenerator #JsoupGenerator
 threads: 16
 index_options:
   - -storePositions
@@ -21,9 +21,9 @@ input: collections/newswire/NYTcorpus/
 index_path: indexes/lucene-index.core17.pos+docvectors+rawdocs # path to the existing index, used in regression test if `--index` option is absent
 collection: NewYorkTimesCollection
 index_stats:
-  documents: 1855649
-  documents (non-empty): 1855649
-  total terms: 751034054
+  documents: 1855650
+  documents (non-empty): 1855650
+  total terms: 751047962
 topics:
   - name: "[TREC 2017 Common Core Track Topics](../src/main/resources/topics-and-qrels/topics.core17.txt)"
     path: topics.core17.txt
@@ -74,9 +74,9 @@ models:
       - -rerankCutoff 20
     results:
       map:
-        - 0.2788
+        - 0.2739
       p30:
-        - 0.4980
+        - 0.4940
   - name: ql
     display: QL
     params:
@@ -105,6 +105,6 @@ models:
       - -rerankCutoff 20
     results:
       map:
-        - 0.2613
+        - 0.2579
       p30:
-        - 0.4953
+        - 0.4893
