+++ src/main/resources/regression/wt10g.yaml
@@ -8,7 +8,7 @@ qrels_root: src/main/resources/topics-and-qrels/
 index_root:
 ranking_root:
 collection: TrecwebCollection
-generator: JsoupGenerator
+generator: DefaultLuceneDocumentGenerator
 threads: 16
 index_options:
   - -storePositions
